feat: add dynamic re-optimization components and services
- Introduced ModalReotimizacao for displaying re-optimization options.
- Added BotaoReotimizar for quick re-optimization actions.
- Implemented AcoesParada for managing actions on specific stops.
- Created AlertaReotimizacao for intelligent alerts based on traffic and delays.
- Developed reotimizacao service with scenarios for dynamic re-optimization.
- Included CSS styles for the new components to ensure proper layout and responsiveness.

+++ speedrota-android/app/src/main/java/br/com/speedrota/ui/screens/dashboard/DashboardScreen.kt
@@ -949,7 +949,7 @@ fun StatusBarPro(item: PieChartItem) {
         }
         Spacer(modifier = Modifier.height(4.dp))
         LinearProgressIndicator(
-            progress = { (item.percent / 100f).coerceIn(0f, 1f) },
+            progress = { (item.percent / 100.0).coerceIn(0.0, 1.0).toFloat() },
             modifier = Modifier
                 .fillMaxWidth()
                 .height(6.dp)
